fix test failes by patching to correct file and delete extra file
